Change: Move useLicense hook to hooks directory
The license feature is unused and could possibly removed completely. But
until now we should get a consistent directory layout.
