+++ src/web/hooks/useLicense.js
